Prevent double escaping forward slash
When subscribing to a topic with a plus in it's topic, it passes through a regex. This regex will replace any occurrences of the "+" sign to `[^\/]*`. 

However, the next rule will escape the forward slash, which was already escaped by the previous `str_replace`, which will produce a warning and on top of that, the topic won't match:

```
PHP Warning:  preg_match(): Unknown modifier ']' in /var/www/vendor/mcfish/libmqtt/src/Client.php on line 586
```

This commit fixes it by providing an unescaped forward slash in the first str_replace, and then allowing the second str_replace to escape the first one.